Created a unified component for all the filters calledFilterBar. Added missing export to GenreFilter.
