+++ src/App.jsx
@@ -1,20 +1,16 @@
-import TestDropdown from "./components/test/TestDropdown";
-import TestFilterContext from "./components/test/TestFilterContext";
-import { FilterProvider } from "./context/FilterContext";
-import TestPlatformFilter from "./components/test/TestPlatformFilter";
-import TestSearchBar from "./components/test/TestSearchBar";
+import TestFilterBar from "./components/test/TestFilterBar";
+
+/* Testing Imports */
+// import TestDropdown from "./components/test/TestDropdown";
+// import TestFilterContext from "./components/test/TestFilterContext";
+// import { FilterProvider } from "./context/FilterContext";
+// import TestPlatformFilter from "./components/test/TestPlatformFilter";
+// import TestSearchBar from "./components/test/TestSearchBar";
 function App() {
   return (
     <>
       <div>Hello World~! go team!</div>
-      <TestDropdown />
-
-      <FilterProvider>
-        <TestFilterContext />
-      </FilterProvider>
-
-      <TestPlatformFilter />
-      <TestSearchBar />
+      <TestFilterBar />
     </>
   );
 }
