+++ src/assets/mockData/gameSessions.js
@@ -1,0 +1,144 @@
+//(Folder Structure for data)
+// src/assets/mockData/gameSessions.js
+
+const gameSessions = [
+  {
+    id: 1,
+    title: "Overwatch 2",
+    description: "Looking for 5man ranked, silver-gold players only",
+    platform: "pc",
+    genres: ["fps"],
+    region: "na",
+    currentPlayers: 2,
+    maxPlayers: 5,
+    status: "ongoing",
+    createdBy: "Player123",
+    customTags: ["mic", "competitive", "dps", "tank"],
+    inviteCode: "OW2-12345",
+  },
+  {
+    id: 2,
+    title: "Valorant",
+    description: "I don't know anyhting about Valorant lel",
+    platform: "pc",
+    genres: ["fps", "strategy"],
+    region: "eu",
+    currentPlayers: 3,
+    maxPlayers: 5,
+    status: "scheduled",
+    scheduledTime: "2025-03-06T15:00:00Z",
+    createdBy: "GamerX",
+    customTags: ["casual", "beginner_friendly", "controller"],
+    inviteCode: "VAL-67890",
+  },
+  {
+    id: 3,
+    title: "Fortnite",
+    description: "Looking for serious players, no trolls please",
+    platform: "ps5",
+    genres: ["battle_royale", "fps"],
+    region: "na",
+    currentPlayers: 1,
+    maxPlayers: 4,
+    status: "ongoing",
+    createdBy: "FortnitePro",
+    customTags: ["competitive", "mic"],
+  },
+  {
+    id: 4,
+    title: "League of Legends",
+    description: "Need support main for our team, lets create that Ocean",
+    platform: "pc",
+    genres: ["moba", "strategy"],
+    region: "eu",
+    currentPlayers: 4,
+    maxPlayers: 5,
+    status: "ongoing",
+    createdBy: "LOLMaster",
+    customTags: ["ranked", "experienced", "healer"],
+  },
+  {
+    id: 5,
+    title: "Minecraft",
+    description: "Creating Your Moms house HAHA!",
+    platform: "pc",
+    genres: ["adventure", "simulation"],
+    region: "na",
+    currentPlayers: 6,
+    maxPlayers: 15,
+    status: "ongoing",
+    createdBy: "BlockBuilder",
+    customTags: ["casual", "beginner_friendly"],
+    inviteCode: "MC-SERVER-12345",
+  },
+  {
+    id: 6,
+    title: "Call of Duty: Warzone",
+    description: "Casual evening games, Call of Dutters AHOY!",
+    platform: "xbox_series",
+    genres: ["fps", "battle_royale"],
+    region: "na",
+    currentPlayers: 2,
+    maxPlayers: 4,
+    status: "scheduled",
+    scheduledTime: "2025-03-05T20:00:00Z",
+    createdBy: "WarzonePlayer",
+    customTags: ["casual", "mic"],
+  },
+  {
+    id: 7,
+    title: "FIFA 25",
+    description: "Looking for all positions except GK",
+    platform: "ps5",
+    genres: ["sports", "simulation"],
+    region: "eu",
+    currentPlayers: 4,
+    maxPlayers: 11,
+    status: "ongoing",
+    createdBy: "FIFALegend",
+    customTags: ["competitive", "mic"],
+  },
+  {
+    id: 8,
+    title: "Among Us",
+    description: "The person reading this might be the imposter (Yoon)",
+    platform: "mobile",
+    genres: ["party", "strategy"],
+    region: "na",
+    currentPlayers: 5,
+    maxPlayers: 10,
+    status: "scheduled",
+    scheduledTime: "2025-03-09T18:00:00Z",
+    createdBy: "SusPlayer",
+    customTags: ["casual", "beginner_friendly"],
+    inviteCode: "AMONGUS-123",
+  },
+  {
+    id: 9,
+    title: "Rocket League",
+    description: "Looking for Diamond+ players for weekend tournament",
+    platform: "pc",
+    genres: ["sports", "racing"],
+    region: "eu",
+    currentPlayers: 2,
+    maxPlayers: 3,
+    status: "ongoing",
+    createdBy: "RocketPro",
+    customTags: ["tournament", "competitive", "mic", "experienced"],
+  },
+  {
+    id: 10,
+    title: "Jackbox Party Pack",
+    description: "Hosting Jackbox Party Pack 5-9, come join!",
+    platform: "pc",
+    genres: ["party"],
+    region: "na",
+    currentPlayers: 2,
+    maxPlayers: 8,
+    status: "ongoing",
+    createdBy: "JackboxHost",
+    customTags: ["casual", "beginner_friendly"],
+  },
+];
+
+export default gameSessions;
